+++ src/components/DatePicker.vue
@@ -1,17 +1,12 @@
 <template>
   <component
-    :is='datePicker'
+    :is='componentName'
     :value='value'
     :from-page.sync='fromPage_'
     :to-page.sync='toPage_'
     :theme-styles='themeStyles_'
-    :drag-attribute='dragAttribute_'
-    :select-attribute='selectAttribute_'
-    :disabled-attribute='disabledAttribute_'
-    :attributes='attributes_'
-    :date-validator='dateValidator'
+    :disabled-attribute='disabledAttribute'
     @drag='dragValue = $event'
-    @input='updateValue'
     v-bind='$attrs'
     v-on='filteredListeners()'
     v-if='isInline'>
@@ -27,28 +22,25 @@
     @didDisappear='popoverDidDisappear'
     v-else>
     <slot
-      :input-value='valueText'
+      :input-value='inputValue'
       :update-value='updateValue'>
       <input
         type='text'
         :class='[inputClass, { "c-input-drag": dragValue }]'
         :style='inputStyle'
-        :placeholder='placeholder_'
-        :value='valueText'
+        :placeholder='inputPlaceholder_'
+        :value='inputValue'
+        :readonly='inputIsReadOnly'
         @change='updateValue($event.target.value)' />
     </slot>
     <component
       slot='popover-content'
-      :is='datePicker'
+      :is='componentName'
       :value='value'
       :from-page.sync='fromPage_'
       :to-page.sync='toPage_'
       :theme-styles='themeStyles_'
-      :drag-attribute='dragAttribute_'
-      :select-attribute='selectAttribute_'
-      :disabled-attribute='disabledAttribute_'
-      :attributes='attributes_'
-      :date-validator='dateValidator'
+      :disabled-attribute='disabledAttribute'
       @drag='dragValue = $event'
       v-bind='$attrs'
       v-on='filteredListeners()'>
@@ -58,18 +50,15 @@
 </template>
 
 <script>
+import Vue from 'vue';
 import Popover from './Popover';
 import SingleDatePicker from './SingleDatePicker';
 import MultipleDatePicker from './MultipleDatePicker';
 import DateRangePicker from './DateRangePicker';
-import DateInfo from '../utils/dateInfo';
+import PickerProfile from '../utils/pickerProfiles';
+import Attribute from '../utils/attribute';
 import defaults from '../utils/defaults';
-import {
-  getDateComps,
-  getNextPage,
-  getMaxPage,
-  getLastArrayItem,
-  blendColors } from '../utils/helpers';
+import { isString } from '../utils/typeCheckers';
 
 export default {
   components: {
@@ -95,87 +84,34 @@ export default {
     dateFormatter: { type: Function, default: defaults.dateFormatter },
     dateParser: { type: Function, default: defaults.dateParser },
     themeStyles: { type: Object, default: () => ({}) },
-    selectColor: { type: String, default: () => defaults.datePickerSelectColor },
-    dragColor: { type: String, default: () => defaults.datePickerDragColor },
-    selectAttribute: Object,
-    dragAttribute: Object,
+    availableDates: Array,
     disabledDates: Array,
-    disabledAttribute: Object,
-    attributes: Array,
   },
   data() {
     return {
       fromPage_: null,
       toPage_: null,
       dragValue: null,
-      valueText: '',
+      inputValue: '',
       popoverForceHidden: false,
+      disablePopoverForceHidden: false,
     };
   },
   computed: {
-    datePicker() {
-      switch (this.mode) {
-        case 'single':
-          return 'single-date-picker';
-        case 'multiple':
-          return 'multiple-date-picker';
-        case 'range':
-          return 'date-range-picker';
-        default:
-          return '';
-      }
+    profile() {
+      return PickerProfile(this.mode, this.dateFormatter, this.dateParser);
     },
-    placeholder_() {
-      if (this.inputPlaceholder) return this.inputPlaceholder;
-      switch (this.mode) {
-        case 'single':
-          return 'Enter Date';
-        case 'multiple':
-          return 'Date 1, Date 2, ...';
-        case 'range':
-          return 'Start Date - End Date';
-        default:
-          return '';
-      }
+    componentName() {
+      return this.profile.componentName;
     },
-    suggestedInputText() {
-      if (!this.value || typeof this.dateFormatter !== 'function') return '';
-      if (this.mode === 'single') {
-        if (typeof this.value.getTime !== 'function') return '';
-        return this.dateFormatter(this.value);
-      } else if (this.mode === 'multiple') {
-        return this.value.length ? this.value.map(d => this.dateFormatter(d)).join(', ') : '';
-      } else if (this.mode === 'range') {
-        if (this.dragValue) {
-          const startText = this.dateFormatter(this.dragValue.start);
-          const endText = this.dateFormatter(this.dragValue.end);
-          return `${startText} - ${endText}`;
-        } else if (this.value) {
-          const startText = this.value.start ? this.dateFormatter(this.value.start) : '';
-          const endText = this.value.end ? this.dateFormatter(this.value.end) : '';
-          if (!startText && !endText) return '';
-          if (!endText) return startText;
-          return `${startText} - ${endText}`;
-        }
-      }
-      return '';
+    inputPlaceholder_() {
+      return this.inputPlaceholder || this.profile.inputPlaceholder;
     },
-    disabledDates_() {
-      if (this.disabledDates) return this.disabledDates.map(d => new DateInfo(d));
-      if (this.disabledAttribute && this.disabledAttribute.dates) return this.disableAttribute.dates.map(d => new DateInfo(d));
-      return [];
+    inputIsReadOnly() {
+      return this.disabledAttribute && this.disabledAttribute.isComplex;
     },
-    dateValidator() {
-      return (date, failEventName) => {
-        if (!this.disabledDates_.length) return true;
-        const dateInfo = date instanceof DateInfo ? date : new DateInfo(date);
-        const disabledDates = this.disabledDates_.filter(d => dateInfo.intersects(d));
-        if (disabledDates && disabledDates.length) {
-          this.$emit(failEventName, disabledDates);
-          return false;
-        }
-        return true;
-      };
+    formattedValue() {
+      return this.profile.formatValue(this.value, this.dragValue);
     },
     themeStyles_() {
       const styles = {
@@ -201,56 +137,13 @@ export default {
         padding: '0',
       };
     },
-    selectAttribute_() {
-      return this.selectAttribute || {
-        highlight: {
-          backgroundColor: this.selectColor,
-          borderWidth: '1px',
-          borderColor: blendColors(this.selectColor, '#000000', 0.1),
-        },
-        contentStyle: {
-          color: '#fafafa',
-        },
-        contentHoverStyle: {
-          backgroundColor: 'transparent',
-          border: '0',
-        },
-      };
-    },
-    dragAttribute_() {
-      return this.dragAttribute || {
-        highlight: {
-          backgroundColor: this.dragColor,
-          height: '25px',
-        },
-        contentHoverStyle: {
-          backgroundColor: 'transparent',
-          border: '0',
-        },
-      };
-    },
-    disabledAttribute_() {
-      const baseAttribute = this.disabledAttribute || {
-        order: 100,
-        contentStyle: {
-          color: '#bcbcbc',
-          textDecoration: 'line-through',
-        },
-        contentHoverStyle: {
-          cursor: 'not-allowed',
-          backgroundColor: 'transparent',
-        },
-      };
-      return {
-        key: 'disabled',
-        ...baseAttribute,
-        dates: this.disabledDates_,
-      };
-    },
-    attributes_() {
-      const attributes = this.attributes ? [...this.attributes] : [];
-      if (this.disabledAttribute_.dates) attributes.push(this.disabledAttribute_);
-      return attributes;
+    disabledAttribute() {
+      if (!this.disabledDates && !this.availableDates) return null;
+      return Attribute({
+        ...defaults.datePickerDisabledAttribute,
+        dates: this.disabledDates,
+        excludeDates: this.availableDates,
+      });
     },
   },
   watch: {
@@ -270,16 +163,23 @@ export default {
       // Clear value on select mode change
       this.$emit('input', null);
     },
-    suggestedInputText(val) {
-      this.valueText = val;
-    },
     value() {
-      if (!this.popoverKeepVisibleOnInput) this.popoverForceHidden = true;
+      this.updateInputValue();
+      if (this.popoverKeepVisibleOnInput) return;
+      if (this.mode === 'multiple') return;
+      if (!this.disablePopoverForceHidden) this.popoverForceHidden = true;
+      this.disablePopoverForceHidden = false;
+    },
+    dragValue() {
+      this.updateInputValue();
+    },
+    disabledAttribute() {
+      this.updateValue(this.value);
     },
   },
   created() {
     this.assignPageRange();
-    this.valueText = this.suggestedInputText;
+    this.updateInputValue();
   },
   methods: {
     filteredListeners() {
@@ -293,55 +193,29 @@ export default {
       this.assignPageRange();
     },
     assignPageRange() {
-      if (!this.value) return;
-      switch (this.mode) {
-        case 'single':
-          this.fromPage_ = getDateComps(this.value);
-          this.toPage_ = getNextPage(this.fromPage_);
-          break;
-        case 'multiple':
-          if (!this.value.length) return;
-          this.fromPage_ = getDateComps(this.value[0]);
-          this.toPage_ = getMaxPage(getDateComps(getLastArrayItem(this.value)), getNextPage(this.fromPage_));
-          break;
-        case 'range':
-          if (!this.value.start || !this.value.end) return;
-          this.fromPage_ = getDateComps(this.value.start);
-          this.toPage_ = getMaxPage(getDateComps(this.value.end), getNextPage(this.fromPage_));
-          break;
-        default:
+      const range = this.profile.getPageRange(this.value);
+      if (range) {
+        this.fromPage_ = range.from;
+        this.toPage_ = range.to;
       }
     },
-    updateValue(value = this.valueText) {
-      if (typeof value === 'string') {
-        this.$emit('input', this.parseValue(value));
+    updateValue(value) {
+      const parsedValue = isString(value) ? this.profile.parseValue(value) : value;
+      // Filter out any disabled dates
+      const filteredValue = this.profile.filterDisabled(parsedValue, this.disabledAttribute);
+      if (this.profile.valuesAreEqual(parsedValue, filteredValue)) {
+        // Everything user entered was accepted so hide popover
+        this.popoverForceHidden = true;
       } else {
-        this.$emit('input', value);
+        // Keep the popover open because something they entered was modified
+        this.disablePopoverForceHidden = true;
       }
+      this.$emit('input', filteredValue);
+      // Update input text for good measure
+      this.updateInputValue();
     },
-    parseValue(valueText) {
-      let value = null;
-      if (this.mode === 'single') {
-        value = this.dateParser(valueText.trim());
-        if (isNaN(value.getTime())) value = null;
-      } else if (this.mode === 'multiple') {
-        value = valueText
-          .split(',')
-          .map(s => this.dateParser(s.trim()))
-          .filter(d => !isNaN(d.getTime()));
-      } else if (this.mode === 'range') {
-        const dates = valueText.split('-').map(s => s.trim());
-        if (dates.length < 2) {
-          value = null;
-        } else {
-          let start = this.dateParser(dates[0]);
-          if (isNaN(start.getTime())) start = null;
-          let end = dates.length > 1 ? this.dateParser(dates[1]) : null;
-          if (end && isNaN(end)) end = null;
-          value = start && end ? { start, end } : null;
-        }
-      }
-      return value;
+    updateInputValue() {
+      Vue.nextTick(() => { this.inputValue = this.profile.formatValue(this.value, this.dragValue); });
     },
   },
 };